requirements.txt
fastapi
uvicorn
langchain==0.3.8
chromadb
sentence-transformers
huggingface-hub
langdetect
beautifulsoup4
requests
langchain-together
pyngrok
nest-asyncio
langchain-community
langchain-chroma
python-dotenv
boto3
langchain-huggingface
streamlit